Add support for unique service principal.
Some JAAS frameworks (e.g. JBoss 5) may create an implicit principal from
username/password credentials passed into the JAAS pipeline. Since the CAS
service URL is used for the username part of the credential, support has
been added to append a unique suffix to the service URL and make it
available to the custom AssertionPrincpal princpal type.

A unqiue URL is generated prior to invoking WebAuthentication#login() to
account for this feature of JBoss 5.